+++ list.h
@@ -22,6 +22,6 @@ int list_for_each(list_t* list, int (*f) (void*, size_t*)); //applies the functi
 void* list_get(list_t*, size_t*, int index); //gets the element at index
 int list_remove(list_t*, void* obj); //remove the first element (???)
 void* list_as_array(list_t*, size_t* size); //returns the list as an array
-int list_length(list_t*); //returns the list length
+size_t list_length(list_t*); //returns the list length
 void list_destroy(list_t*); //free the memory used by the list
 #endif //SOP_SYNTHESIS_LIST_H
